Working on testing system implementation 0.4

+++ src/main/java/com/example/springexercise3boot/repositories/TestsRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TestsRepository extends JpaRepository<Test, Integer> {
-
+public interface TestsRepository extends JpaRepository<Test, Long> {
 }
